nicer intf for getting constraint system
