+++ src/lib/core.ts
@@ -1,4 +1,4 @@
-import { Bool, Field, Scalar, Group } from '../snarky.js';
+import { Bool, Field, Scalar, Group, Keypair } from '../snarky.js';
 
 export { Field, Bool };
